+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 ### Added
 * Added fetal medicine and dating options to the test group filter (#50)
 * Added new gestation at delivery and booking date to case field existance filter (#52)
+
+### Changed
+* Changed the values of registry to use registry name rather than codes and added new supra-region options (#54)
 ## 4.4.1 / 2021-01-13
 ### Fixed
 * Relax constraints to support Rails 6.x