+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+*no unreleased changes*
+## 5.0.0 / 2021-01-19
 ### Added
 * Added fetal medicine and dating options to the test group filter (#50)
 * Added new gestation at delivery and booking date to case field existance filter (#52)